requirements.txt
alembic==1.14.1
altair==5.5.0
amqp==5.3.1
attrs==25.1.0
Automat==24.8.1
beautifulsoup4==4.13.3
billiard==4.2.1
blinker==1.9.0
branca==0.8.1
cachetools==5.5.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.1
cssselect==1.2.0
cycler==0.12.1
dash==2.9.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
defusedxml==0.7.1
dnspython==2.7.0
email_validator==2.2.0
filelock==3.17.0
Flask==3.1.0
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
folium==0.19.4
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
kombu==5.4.2
lxml==5.3.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mysqlclient==2.2.7
narwhals==1.26.0
numpy==2.2.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parsel==1.10.0
pillow==11.1.0
pip==25.0.1
plotly==6.0.0
prompt_toolkit==3.0.50
Protego==0.4.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydeck==0.9.1
PyDispatcher==2.0.7
Pygments==2.19.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
queuelib==1.7.0
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
rpds-py==0.22.3
scikit-learn==1.6.1
scipy==1.15.1
Scrapy==2.12.0
selenium==4.30.0
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.38
streamlit==1.42.0
tenacity==9.0.0
threadpoolctl==3.5.0
tldextract==5.1.3
toml==0.10.2
tornado==6.4.2
trio==0.29.0
trio-websocket==0.12.2
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
w3lib==2.3.1
waitress==3.0.2
watchdog==6.0.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
xyzservices==2025.1.0
zope.interface==7.1.1